feat(erc2981Settable): tests

+++ interfaces/IERC2981Settable.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.0;
 import "openzeppelin/contracts/interfaces/IERC2981.sol";
 
 /**
-* @title ERC-2981 royalty setter
+* @title ERC-2981 royalty setter interface
 * @dev Custom implementation, ERC-2981 does not include royalty settings.
 */
 interface IERC2981Settable is IERC2981 {
@@ -12,7 +12,7 @@ interface IERC2981Settable is IERC2981 {
     * @param recipient The receiver of royalty
     * @param royaltyPercent The royalty percentage (using 2 decimals - 10000 = 100%, 0 = 0%)
     */
-    function setDefaultRoyalty(address recipient, uint16 royaltyPercent) external;
+    function setDefaultRoyalty(address recipient, uint96 royaltyPercent) external;
 
     /**
     * @notice Set royalty for a token
@@ -20,7 +20,7 @@ interface IERC2981Settable is IERC2981 {
     * @param recipient The receiver of royalty
     * @param royaltyPercent The royalty percentage (using 2 decimals - 10000 = 100%, 0 = 0%)
     */
-    function setTokenRoyalty(uint256 tokenId, address recipient, uint16 royaltyPercent) external;
+    function setTokenRoyalty(uint256 tokenId, address recipient, uint96 royaltyPercent) external;
 
     /**
     * @notice Update royalty recipient for whole collection
